requirements.txt
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
